ENter

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Permission\Models\Role;
 use App\Permission\Models\Permission;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -15,6 +16,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('haveaccess', 'role.index');
+
         $roles = Role::orderBy('id', 'Desc')->paginate(10);
 
         return view('role.index', compact('roles'));
@@ -27,6 +30,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('haveaccess', 'role.create');
+
         $permissions = Permission::get();
         return view('role.create', compact('permissions'));
     }
@@ -39,6 +44,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('haveaccess', 'role.create');
+
         $request->validate([
             'name'        => 'required|max:50|unique:roles,name',
             'slug'        => 'required|max:50|unique:roles,slug',
@@ -64,6 +71,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+
+        $this->authorize('haveaccess', 'role.show');
         $permissions_role = [];
 
         foreach ($role->permissions as $permission) {
@@ -82,6 +91,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('haveaccess', 'role.edit');
         $permissions_role = [];
 
         foreach ($role->permissions as $permission) {
@@ -101,6 +111,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        $this->authorize('haveaccess', 'role.edit');
+
         $request->validate([
             'name'        => 'required|max:50|unique:roles,name,'.$role->id,
             'slug'        => 'required|max:50|unique:roles,slug,'.$role->id,
@@ -125,6 +137,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('haveaccess', 'role.destroy');
+
         $role->delete();
         return redirect()->route('role.index')
                 ->with('status_success', 'Role successfully removed');
